@@ -3,7 +3,8 @@
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
+import { defineComponent } from 'vue';
+import Vue, { compile } from 'vue';
 
 const alias = {
   // algemeen
@@ -72,7 +73,7 @@ const alias = {
   rechten: 'key',
 };
 
-export default Vue.extend({
+export default defineComponent({
   props: {
     icon: {
       default: '',
