@@ -10,10 +10,11 @@
 </template>
 
 <script lang="ts">
-import Vue, { PropType } from 'vue';
+import type { PropType } from 'vue';
+import { defineComponent } from 'vue';
 import { KeuzeOptie } from '../../model/groep';
 
-export default Vue.extend({
+export default defineComponent({
   props: {
     keuzes: {
       required: true,
